Implemented validation settings

+++ src/main/java/files/route/dijkstra/ConfigFileCreator.java
@@ -1,5 +1,6 @@
 package files.route.dijkstra;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -14,22 +15,26 @@ public class ConfigFileCreator {
         String absolutePath = "";
 
         try {
-            Files.createDirectories(Path.of(diretorioConfiguration));
+            if(!new File(diretorioConfiguration).exists()){
+                Files.createDirectories(Path.of(diretorioConfiguration));
 
-            Path arquivoConfigPath = Path.of(diretorioConfiguration, arquivoConfig);
-            StringBuilder conteudoArquivo = new StringBuilder();
-            conteudoArquivo
-                    .append("processed=")
-                    .append(diretorioTest)
-                    .append("/processed\n")
-                    .append("unprocessed=")
-                    .append(diretorioTest)
-                    .append("/unprocessed\n");
+                Path arquivoConfigPath = Path.of(diretorioConfiguration, arquivoConfig);
+                StringBuilder conteudoArquivo = new StringBuilder();
+                conteudoArquivo
+                        .append("processed=")
+                        .append(diretorioTest)
+                        .append("/processed\n")
+                        .append("unprocessed=")
+                        .append(diretorioTest)
+                        .append("/unprocessed\n");
 
-            absolutePath = arquivoConfigPath.toAbsolutePath().toString();
+                absolutePath = arquivoConfigPath.toAbsolutePath().toString();
 
-            Files.write(arquivoConfigPath, conteudoArquivo.toString().getBytes(), StandardOpenOption.CREATE);
-            System.out.println("Configuration file created: " + arquivoConfigPath.toAbsolutePath());
+                Files.write(arquivoConfigPath, conteudoArquivo.toString().getBytes(), StandardOpenOption.CREATE);
+                System.out.println("Configuration file created: " + arquivoConfigPath.toAbsolutePath());
+            } else {
+                absolutePath = Path.of(diretorioConfiguration, arquivoConfig).toString();
+            }
         } catch (IOException ex) {
             System.out.println("Error creating directories or writing configuration file: " + ex.getMessage());
         }
